+++ lib/src/query_executors/postgre_sql_executor.dart
@@ -27,10 +27,9 @@ class PostgreSqlExecutor implements QueryExecutor {
   }
 
   @override
-  Future<List<List>> query(
-      String query, Map<String, dynamic> substitutionValues,
-      [List<String> returningFields]) {
-    if (returningFields != null) {
+  Future<List<List>> query(String query, Map<String, dynamic> substitutionValues, [List<String> returningFields]) {
+    if (returningFields?.isNotEmpty == true) {
+    //if (returningFields != null) {
       var fields = returningFields.join(', ');
       var returning = 'RETURNING $fields';
       query = '$query $returning';
@@ -38,19 +37,18 @@ class PostgreSqlExecutor implements QueryExecutor {
 
     logger?.fine('Query: $query');
     logger?.fine('Values: $substitutionValues');
-    //print('Query ${query}');
-    //print('Values ${substitutionValues}');
+    //print('Query: $query');
+    //print('Values: $substitutionValues');
+    //print('Fields: $returningFields');
 
     return _connection.query(query, substitutionValues: substitutionValues);
   }
 
   @override
-  Future<List<Map<String, dynamic>>> getAsMap(String query,
-      {Map<String, dynamic> substitutionValues}) async {
+  Future<List<Map<String, dynamic>>> getAsMap(String query, {Map<String, dynamic> substitutionValues}) async {
     //print('PostgreSqlExecutor@getAsMap query $query');
     //print('PostgreSqlExecutor@getAsMap substitutionValues $substitutionValues');
-    var rows =
-        await getAsMapWithMeta(query, substitutionValues: substitutionValues);
+    var rows = await getAsMapWithMeta(query, substitutionValues: substitutionValues);
 
     final result = <Map<String, dynamic>>[];
     if (rows != null || rows.isNotEmpty) {
@@ -73,14 +71,12 @@ class PostgreSqlExecutor implements QueryExecutor {
       {Map<String, dynamic> substitutionValues}) {
     logger?.fine('Query: $query');
     logger?.fine('Values: $substitutionValues');
-    return _connection.mappedResultsQuery(query,
-        substitutionValues: substitutionValues);
+    return _connection.mappedResultsQuery(query, substitutionValues: substitutionValues);
   }
 
   //Use generic function type syntax for parameters.
   //Future<dynamic> f(QueryExecutor connection)
-  Future<dynamic> simpleTransaction(
-      Future<dynamic> Function(QueryExecutor) f) async {
+  Future<dynamic> simpleTransaction(Future<dynamic> Function(QueryExecutor) f) async {
     logger?.fine('Entering simpleTransaction');
     if (_connection is! PostgreSQLConnection) {
       return await f(this);
@@ -106,8 +102,7 @@ class PostgreSqlExecutor implements QueryExecutor {
       if (txResult.reason == null) {
         throw StateError('The transaction was cancelled.');
       } else {
-        throw StateError(
-            'The transaction was cancelled with reason "${txResult.reason}".');
+        throw StateError('The transaction was cancelled with reason "${txResult.reason}".');
       }
     } else {
       return returnValue;
@@ -138,20 +133,12 @@ class PostgreSqlExecutor implements QueryExecutor {
       if (txResult.reason == null) {
         throw StateError('The transaction was cancelled.');
       } else {
-        throw StateError(
-            'The transaction was cancelled with reason "${txResult.reason}".');
+        throw StateError('The transaction was cancelled with reason "${txResult.reason}".');
       }
     } else {
       return returnValue;
     }
   }
-
-  /*@override
-  Future transaction2(Function queryBlock, {int commitTimeoutInSeconds}) {
-    var conn = _connection as PostgreSQLConnection;
-    return conn.transaction(queryBlock,
-        commitTimeoutInSeconds: commitTimeoutInSeconds);
-  }*/
 }
 
 /// A [QueryExecutor] that manages a pool of PostgreSQL connections.
@@ -172,9 +159,7 @@ class PostgreSqlExecutorPool implements QueryExecutor {
   final Pool _pool, _connMutex = Pool(1);
   List<String> schemes = ['public'];
 
-  PostgreSqlExecutorPool(this.size, this.connectionFactory,
-      {this.logger, this.schemes})
-      : _pool = Pool(size) {
+  PostgreSqlExecutorPool(this.size, this.connectionFactory, {this.logger, this.schemes}) : _pool = Pool(size) {
     assert(size > 0, 'Connection pool cannot be empty.');
   }
 
@@ -193,9 +178,7 @@ class PostgreSqlExecutorPool implements QueryExecutor {
           logger?.fine('Spawning connections...');
           final conn = connectionFactory();
 
-          final executor = await conn
-              .open()
-              .then((_) => PostgreSqlExecutor(conn, logger: logger));
+          final executor = await conn.open().then((_) => PostgreSqlExecutor(conn, logger: logger));
 
           //isso executa uma query para definir os esquemas
           if (schemes != null && schemes.isNotEmpty) {
@@ -220,17 +203,16 @@ class PostgreSqlExecutorPool implements QueryExecutor {
 
   @override
   Future<List<Map<String, Map<String, dynamic>>>> getAsMapWithMeta(String query,
-      {Map<String, dynamic> substitutionValues}) {
+      {Map<String, dynamic> substitutionValues, List<String> returningFields}) {
     return _pool.withResource(() async {
       final executor = await _next();
-      return executor.getAsMapWithMeta(query,
-          substitutionValues: substitutionValues);
+      return executor.getAsMapWithMeta(query, substitutionValues: substitutionValues);
     });
   }
 
   @override
   Future<List<Map<String, dynamic>>> getAsMap(String query,
-      {Map<String, dynamic> substitutionValues}) async {
+      {Map<String, dynamic> substitutionValues, returningFields}) async {
     return _pool.withResource(() async {
       final executor = await _next();
       return executor.getAsMap(query, substitutionValues: substitutionValues);
@@ -245,9 +227,7 @@ class PostgreSqlExecutorPool implements QueryExecutor {
   }
 
   @override
-  Future<List<List>> query(
-      String query, Map<String, dynamic> substitutionValues,
-      [List<String> returningFields]) {
+  Future<List<List>> query(String query, Map<String, dynamic> substitutionValues, [List<String> returningFields]) {
     return _pool.withResource(() async {
       final executor = await _next();
       return executor.query(query, substitutionValues, returningFields);
@@ -261,11 +241,4 @@ class PostgreSqlExecutorPool implements QueryExecutor {
       return executor.transaction(f);
     });
   }
-
- /* @override
-  Future transaction2(Function queryBlock, {int commitTimeoutInSeconds}) async {
-    var executor = await _next();
-    return executor.transaction2(queryBlock,
-        commitTimeoutInSeconds: commitTimeoutInSeconds);
-  }*/
 }