+++ CHANGELOG.md
@@ -33,3 +33,11 @@
 ## 1.1.2
 
 - add orWhereGroup and whereGroup method for grouping conditions
+
+## 1.1.8
+
+- implemented insertGetId and insertGetAll methods for returning id and row post inserted
+
+## 1.1.9
+
+- implemented setAll method for insert and update